WIP: mm/vas: Attempt to fix MVAS for multi-threaded applications
At the moment multi-threaded applications can confuse the kernel with the
sharing of the underlying memory regions of MVAS. Parts of the memory
regions might be overwritten and get lost by conflicting updates which can
cause segmentation faults or bus faults of the application.

At the moment this commit is only a partial fix of the underlying problem.
I am still working on fixing it completely.

Signed-off-by: Till Smejkal <till.smejkal@gmail.com>